refactor: move useWallet hook back into 'features/wallet'
